requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
cinemagoer==2023.5.1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.1.1
django-browser-reload==1.16.0
django-tailwind==3.8.0
djhtml==3.0.6
djlint==1.35.2
EditorConfig==0.12.4
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httplib2==0.22.0
idna==3.10
IMDbPY==2022.7.9
jsbeautifier==1.15.1
json5==0.9.25
lxml==5.3.0
oauthlib==3.2.2
pathspec==0.12.1
pillow==10.4.0
PlexAPI==4.15.16
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-cache==0.5.2
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
tmdbsimple==2.9.1
tqdm==4.66.5
trakt==3.4.0
tvdb_api==3.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
